refactor: centralize data enhancer callbacks
